# stat/distuv: improve Pareto.Rand performance, also adding Pareto.Rand benchmark