Please take a look.

Just removes an unnecessary math.Log call.

benchstat:
```
name        old time/op  new time/op  delta
ParetoRand  38.0ns ± 8%  24.4ns ± 2%  -35.72%  (p=0.000 n=9+10)
```